move form-related classes to their own namespace
