+++ src/AdminPages/AdminController.php
@@ -4,6 +4,7 @@ namespace Ecurring\WooEcurring\AdminPages;
 
 use ChriCo\Fields\ViewFactory;
 use Dhii\Output\Template\TemplateInterface;
+use Ecurring\WooEcurring\AdminPages\Form\FormFieldsCollectionBuilder;
 use Ecurring\WooEcurring\Settings\SettingsCrudInterface;
 use eCurring_WC_Plugin;
 use Throwable;
